requirements.txt
openai
requests
python-dotenv
Flask
flask-cors>=3.0.10
